+++ src/protocols/gamespy/protocols/mod.rs
@@ -1,5 +1,2 @@
 pub mod one;
 pub mod three;
-
-pub use one::*;
-pub use three::*;