+++ src/protocols/gamespy/protocols/three/protocol.rs
@@ -1,7 +1,6 @@
 use crate::bufferer::{Bufferer, Endianess};
 use crate::protocols::gamespy::common::has_password;
-use crate::protocols::gamespy::three::{Player, Response};
-use crate::protocols::gamespy::Team;
+use crate::protocols::gamespy::three::{Player, Response, Team};
 use crate::protocols::types::TimeoutSettings;
 use crate::socket::{Socket, UdpSocket};
 use crate::{GDError, GDResult};