+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Services:
 Protocols:
 - Valve: Request type enums have been renamed from all caps to starting-only uppercase, ex: `INFO` to `Info`
 - GameSpy 1: `players_minimum` is now an `Option<u8>` instead of an `u8`
+- GameSpy 1: Is now under `protocols::gamespy::one` instead of `protocols::gamespy`
 
 # 0.2.1 - 03/03/2023
 ### Changes: